# doc/api.yaml
openapi: 3.1.0
info:
  title: Wasapp Service API Documentation
  description: |-
    API for interacting with Wasapp, a messaging service.
  version: 1.0.0
servers:
  - url: https://wasapp.online/v1

# ================================================================
...
paths:
...
  /users:
    get:
      tags: ["Search"]
      summary: List or search users on the platform
      description: |-
        Returns a list of users.
          - If a 'username' query parameter is provided, returns users whose usernames most closely match the query (ordered by similarity).
          - Otherwise, returns all users, ordered alphabetically by username.
      operationId: userSearch
      parameters:
        - name: username
          in: query
          description: Filter users whose username is similar to the provided value.
          required: false
          schema:
            $ref: '#/components/schemas/UserName'
      responses:
        "200":
          description: Users list returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersSearchResponse'

  # -------------------- USERNAME UPDATE --------------------
  /me/name:
    patch:
      tags: ["User Management"]
      summary: Update the username of the logged-in user
      description: Updates the username of the logged-in user using the new value provided in the request body.
      operationId: setMyUserName
      security: [ { BearerAuth: [] } ]
      requestBody:
        description: The new username for the logged-in user.
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/PatchUserNameRequest'
      responses:
        "200":
          description: Username updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Invalid username (too short or too long or invalid chars).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authentication required or failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Username already in use.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

# -------------------- USER AVATAR IMAGE --------------------
  /me/avatar:
    put:
      tags: ["User Management"]
      summary: Upload or replace the avatar of the logged-in user
      description: >-
        Uploads or replaces the avatar for the authenticated user.
        The image must be JPEG, PNG or WebP format, maximum 5MB.
        Returns the updated user object with the new image URL.
      operationId: setMyUserAvatar
      security: [ { BearerAuth: [] } ]
      requestBody:
        description: Image file for user avatar
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BinaryImageUploadRequest'
            encoding:
              binaryImage:
                contentType: image/jpeg, image/png, image/webp
      responses:
        "204":
          description: Avatar uploaded successfully. No content is returned.
        "400":
          description: Bad request - invalid image format or too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authentication required or failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "413":
          description: Image file exceeds maximum size (5MB).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "415":
          description: Unsupported image format. Only JPEG, PNG, WebP allowed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # -------------------- CREATE CHATS --------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "JWT"

  # ================================================================
  schemas:
...
    BaseId:
      type: string
      format: uuid
      readOnly: true
      description: Base UUID identifier type, used only as a parent for specific ID schemas
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      minLength: 36
      maxLength: 36
      example: "550e8400-e29b-41d4-a716-446655440000"
    BaseName:
      description: >-
        The name must be 3 to 24 characters long.
        Allowed characters are letters, numbers, spaces, underscores (_), and hyphens (-).
        The name needs at least one alpha-numeric character.
      type: string
      pattern: "^(?=.*[A-Za-z0-9_-])[A-Za-z0-9 _-]{3,24}$"
      minLength: 3
      maxLength: 24
...
    BinaryImage:
      type: string
      format: binary
      minLength: 1
      maxLength: 5242880
      description: Image file (JPEG, PNG, WebP, max 5MB), used for uploading images.

    # -------------------- USER --------------------
    User:
      type: object
      description: User information.
      required: [id, name]
      properties:
        id:     { $ref: '#/components/schemas/UserId'   }
        name:   { $ref: '#/components/schemas/UserName' }
    UserId:
      allOf:
        - $ref: '#/components/schemas/BaseId'
        - description: Unique identifier assigned to the user by the server.
    UserName:
      allOf:
        - $ref: '#/components/schemas/BaseName'
        - description: Unique username visible to other users.
        - example: Feldspar

    # -------------------- LOGIN --------------------
...
    PatchUserNameRequest:
      type: object
      description: Request body for updating the logged-in user's username.
      properties:
        userName: { $ref: '#/components/schemas/UserName' }
...
    BinaryImageUploadRequest:
      type: object
      description: Request for uploading a image file (JPEG, PNG, WebP, max 5MB), used for uploading images.
      required: [binaryImage]
      properties:
        binaryImage: { $ref: '#/components/schemas/BinaryImage' }

    # -------------------- RESPONSES --------------------
    Error:
      type: object
      description: Default error structure.
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code for frontend logic.
          enum:
            - CHAT_NOT_FOUND
            - USER_NOT_FOUND
            - ALREADY_IN_GROUP
            - NOT_A_GROUP_CHAT
            - FORBIDDEN_NOT_MEMBER
            - INVALID_INPUT
            - INTERNAL_SERVER_ERROR
            - UNAUTHORIZED
            - USERNAME_TAKEN
            - CHAT_ALREADY_EXISTS
            - PAYLOAD_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
          example: USER_NOT_FOUND
        message:
          type: string
          description: Human-readable error message.
          example: The specified user ID does not exist.
...
    UsersSearchResponse:  #todo: paging
      type: object
      description: Response containing the list of users matching the search criteria.
      required: [usersList]
      properties:
        usersList:
          type: array
          description: >-
            Returns the list of user matching the username query (ordered by similarity),
            or all users alphabetically if no query is provided.
          minItems: 0
          maxItems: 50
          items:
            $ref: '#/components/schemas/User'
...